+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/service/timesheet/TimesheetServiceImpl.java
@@ -538,7 +538,8 @@ public class TimesheetServiceImpl implements TimesheetService {
             .convert(
                 appHumanResourceService.getAppBase().getUnitHours(),
                 product.getUnit(),
-                hoursDuration);
+                hoursDuration,
+                AppBaseService.DEFAULT_NB_DECIMAL_DIGITS);
 
     PriceList priceList =
         Beans.get(PartnerPriceListService.class)