+++ axelor-production/src/main/java/com/axelor/apps/production/service/CostSheetServiceImpl.java
@@ -304,7 +304,11 @@ public class CostSheetServiceImpl implements CostSheetService {
       Product product = prodHumanResource.getProduct();
 
       costPerHour =
-          unitConversionService.convert(hourUnit, product.getUnit(), product.getCostPrice());
+          unitConversionService.convert(
+              hourUnit,
+              product.getUnit(),
+              product.getCostPrice(),
+              appProductionService.getNbDecimalDigitForUnitPrice());
     }
 
     BigDecimal durationHours =
@@ -514,7 +518,11 @@ public class CostSheetServiceImpl implements CostSheetService {
     if (prodHumanResource.getProduct() != null) {
       Product product = prodHumanResource.getProduct();
       costPerHour =
-          unitConversionService.convert(hourUnit, product.getUnit(), product.getCostPrice());
+          unitConversionService.convert(
+              hourUnit,
+              product.getUnit(),
+              product.getCostPrice(),
+              appProductionService.getNbDecimalDigitForUnitPrice());
     }
     BigDecimal durationHours =
         new BigDecimal(realDuration)