+++ CHANGELOG.md
@@ -87,6 +87,7 @@ When we update an existing move in daybook mode, we update the partner balances 
 - PARTNER : contact partner form view If the option generatePartnerSequence is false in Base App,allow to edit the field partnerSeq.
 - MANUF ORDER : Exception Message tracking number not filled
 - ICalendar: Fix Nullpointer Exception
+- Fixed occasionnal scale rounding anomalies on unit conversions.
 
 
 ## [5.0.6] - 2018-10-06